requirements.txt
streamlit
langchain
langchain-google-genai
langchain-pinecone
pypdf2
pinecone-client
sentence-transformers
langchain-community
python-dotenv
numpy